deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy==1.26.4
scikit-learn==1.5.0
Pillow
plotly